+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 * fix bug if secret not correct, server raise error and exit
 * fix cli client usage, options should should in front of the url param 
+* bug fix : begin request url: port=4821 server stops
 
 ### 0.5.2
 
 * http get support headers
+* bug fix ensureRequest's retry
 
 ### 0.5.0